pyproject.toml
[tool.poetry]
name = "juntos-somos-mais"
version = "0.0.1"
description = "Juntos Somos Mais Api"
authors = ["RempelOliveira <rempel.oliveira@gmail.com>"]
readme = "README.md"
packages = [{include = "api"}]

[tool.poetry.dependencies]
python = "^3.12.3"
python-jose = "^3.3.0"
fastapi = "^0.111.0"
mangum = "^0.17.0"
pyjwt = "^2.8.0"
prettyconf = "^2.2.1"
requests = "^2.32.2"
inflection = "^0.5.1"

[tool.poetry.group.dev.dependencies]
faker = "^25.2.0"
pytest = "^8.2.1"
pytest-cov = "^5.0.0"
ruff = "^0.4.5"
isort = "^5.13.2"

[tool.pytest.ini_options]
pythonpath = ["."]
python_files = ["*.py"]
python_classes = ["*UnitTests", "*IntegrationTests"]
python_functions = ["*Test"]
addopts = "--tb=native --cov-report=term-missing --cov=."
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.report]
omit = ["tests/*", "*/__init__.py"]
fail_under = 80

[tool.ruff]
select = ["ALL"]
tab-size = 4
line-length = 127
ignore = [
    "D",
    "I",
    "B004",
    "B009",
    "N802",
    "E722",
    "FA100",
    "FA102",
    "PT009",
    "PT013",
    "ANN101",
    "BLE001",
    "COM812",
    "TCH001",
    "PLW2901"
]

[tool.ruff.extend-per-file-ignores]
"tests/**/*.py" = [
    "S101"
]

[tool.ruff.pylint]
max-args = 3
max-returns = 4

[tool.isort]
indent = 4
line_length = 127
multi_line_output = 2
include_trailing_comma = false
force_grid_wrap = 0
ensure_newline_before_comments = true
remove_redundant_aliases = true
lines_between_types = 1
lines_after_imports = 2
length_sort = true
src_paths = [
    "api",
    "tests"
]

sections = [
    "THIRDPARTY",
    "FIRSTPARTY"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
